pyproject.toml
[project]
name = "wealthboard-hexagonal"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "yfinance",
    "pandas",
    "oracledb",
    "python-dotenv"
]


	
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"
	
[tool.setuptools.packages.find]
where = ["src"]

#[project.scripts]
# tbd